refactor: move download-section and tooltip wrapper
Move download-section from src/components/ to src/components/shared/.
Move coming-soon-tooltip-wrapper from src/components/ to src/components/ui-custom/.

- Update imports in batch views, tools, pricing, and subtitle-translator-main
- Add type safety to generateContent in subtitle-translator-main
